requirements.txt
flask
flask-cors
replicate
openai==0.28
requests
Pillow